requirements.txt
fastapi
uvicorn
openai
chromadb
tiktoken
python-dotenv
groq
requests
python-docx 
pypdf
python-dotenv
python-multipart
pymongo